chore: Add OutputStreamWriterExample class to write data to the output stream
This commit adds the OutputStreamWriterExample class, which demonstrates how to use OutputStreamWriter to write data to the output stream. It includes creating an instance of OutputStreamWriter, writing data to the output stream, and reading data from the input stream. This addition enhances the functionality of the code by providing an example of using OutputStreamWriter for writing text data to the output stream.