chore: Add FileWriterExample class to write a text file using FileWriter
This commit adds the FileWriterExample class, which demonstrates how to write a text file using FileWriter. It includes creating a FileWriter object, writing a sample text file, and flushing the writer. This addition enhances the functionality of the code by providing an example of using FileWriter for writing text files.